@@ -1,49 +1,46 @@
-import React, { Component } from "react";
+import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
 import PropTypes from "prop-types";
 import Selector from "./Selector";
 import * as BooksAPI from "./BooksAPI";
 
-class BookPage extends Component {
-  state = {
-    book: {},
-    isLoaded: false,
-  };
-  async componentDidMount() {
-    const book = await BooksAPI.get(this.props.match.params.bookId);
-    this.setState({ isLoaded: true, book: book });
-  }
+export default function BookPage(props) {
+  const { shelves, onUpdateBook } = props;
 
-  render() {
-    console.log(this.props.location.state.pathname);
-    const { shelves, onUpdateBook } = this.props;
-    const { book } = this.state;
-    if (!this.state.isLoaded) {
-      return <div>Loading...</div>;
-    } else {
-      return (
-        <div className="book-page">
-          <div className="book-header">
-            <img src={book.imageLinks.thumbnail} alt="book cover" />
-            <div className="book-card">
-              <h1>{book.title}</h1>
-              <h4>{book.subtitle}</h4>
-              <h3>{book.authors && book.authors.join(", ")}</h3>
-              <p>
-                {book.publisher}, {book.publishedDate}
-              </p>
-              <p>Pages: {book.pageCount}</p>
-            </div>
-            <Link to={this.props.location.state.pathname}>
-              <button className="close-search">Close</button>
-            </Link>
+  const [book, setBook] = useState({});
+  const [isLoaded, setIsLoaded] = useState(false);
+
+  useEffect(async () => {
+    const book = await BooksAPI.get(props.match.params.bookId);
+    setBook(book);
+    setIsLoaded(true);
+  });
+
+  if (!isLoaded) {
+    return <div>Loading...</div>;
+  } else {
+    return (
+      <div className="book-page">
+        <div className="book-header">
+          <img src={book.imageLinks.thumbnail} alt="book cover" />
+          <div className="book-card">
+            <h1>{book.title}</h1>
+            <h4>{book.subtitle}</h4>
+            <h3>{book.authors && book.authors.join(", ")}</h3>
+            <p>
+              {book.publisher}, {book.publishedDate}
+            </p>
+            <p>Pages: {book.pageCount}</p>
           </div>
-          <p>{book.description}</p>
-          <a href={book.infoLink}>More about {book.title}</a>
-          <Selector book={book} shelves={shelves} onUpdateBook={onUpdateBook} />
+          <Link to={props.location.state.pathname}>
+            <button className="close-search">Close</button>
+          </Link>
         </div>
-      );
-    }
+        <p>{book.description}</p>
+        <a href={book.infoLink}>More about {book.title}</a>
+        <Selector book={book} shelves={shelves} onUpdateBook={onUpdateBook} />
+      </div>
+    );
   }
 }
 
@@ -51,5 +48,3 @@ BookPage.propTypes = {
   shelves: PropTypes.array.isRequired,
   onUpdateBook: PropTypes.func.isRequired,
 };
-
-export default BookPage;
