@@ -1,48 +1,48 @@
-import React, { Component } from "react";
+import React, { useState } from "react";
 import PropTypes from "prop-types";
 import SearchInput from "./SearchInput";
 import SearchDisplay from "./SearchDisplay";
 import * as BooksAPI from "./BooksAPI";
 
-class SearchPage extends Component {
-  state = {
-    query: "",
-    foundBooks: [],
-    invalidQuery: false,
-  };
+export default function SearchPage(props) {
+  const { books, shelves, onUpdateBook } = props;
 
-  updateQuery = (e) => {
+  const [query, setQuery] = useState("");
+  const [foundBooks, setFoundBooks] = useState([]);
+  const [invalidQuery, setInvalidQuery] = useState(false);
+
+  const updateQuery = (e) => {
     const query = e.target.value;
-    this.setState({ query: query });
+    setQuery(query);
     if (query === "") {
-      this.setState({ foundBooks: [] });
+      setFoundBooks([]);
       return;
     }
     BooksAPI.search(query).then((books) => {
-      books.error
-        ? this.setState({ foundBooks: [], invalidQuery: true })
-        : this.setState({ foundBooks: books, invalidQuery: false });
+      if (books.error) {
+        setFoundBooks([]);
+        setInvalidQuery(true);
+      } else {
+        setFoundBooks(books);
+        setInvalidQuery(false);
+      }
     });
   };
 
-  render() {
-    const { query, foundBooks, invalidQuery } = this.state;
-    const { books, shelves, onUpdateBook } = this.props;
-    return (
-      <div className="SearchPage">
-        <SearchInput query={query} onInputChange={this.updateQuery} />
-        {foundBooks && foundBooks.length > 0 && (
-          <SearchDisplay
-            books={books}
-            foundBooks={foundBooks}
-            shelves={shelves}
-            onUpdateBook={onUpdateBook}
-          />
-        )}
-        <p>{invalidQuery ? "Invalid query" : null}</p>
-      </div>
-    );
-  }
+  return (
+    <div className="SearchPage">
+      <SearchInput query={query} onInputChange={updateQuery} />
+      {foundBooks && foundBooks.length > 0 && (
+        <SearchDisplay
+          books={books}
+          foundBooks={foundBooks}
+          shelves={shelves}
+          onUpdateBook={onUpdateBook}
+        />
+      )}
+      <p>{invalidQuery ? "Invalid query" : null}</p>
+    </div>
+  );
 }
 
 SearchPage.propTypes = {
@@ -50,5 +50,3 @@ SearchPage.propTypes = {
   shelves: PropTypes.array.isRequired,
   onUpdateBook: PropTypes.func.isRequired,
 };
-
-export default SearchPage;
