@@ -1,32 +1,30 @@
-import React, { Component } from "react";
+import React from "react";
 import PropTypes from "prop-types";
 
-class Selector extends Component {
-  updateShelf = (e) => {
-    const book = this.props.book;
+export default function Selector(props) {
+  const { book, shelves, onUpdateBook } = props;
+
+  const updateShelf = (e) => {
     book.shelf = e.target.value;
-    this.props.onUpdateBook(book);
+    onUpdateBook(book);
   };
 
-  render() {
-    const { book, shelves } = this.props;
-    return (
-      //TODO: Refactor selected options in DRY manner (loop through shelves?)
-      <div className="book-shelf-changer">
-        <select value={book.shelf || "none"} onChange={(e) => this.updateShelf(e)}>
-          <option value="move" disabled>
-            Move to...
+  return (
+    //TODO: Refactor selected options in DRY manner (loop through shelves?)
+    <div className="book-shelf-changer">
+      <select value={book.shelf || "none"} onChange={(e) => updateShelf(e)}>
+        <option value="move" disabled>
+          Move to...
+        </option>
+        {shelves.map((shelf) => (
+          <option key={shelf.name} value={shelf.name}>
+            {shelf.displayName}
           </option>
-          {shelves.map((shelf) => (
-            <option key={shelf.name} value={shelf.name}>
-              {shelf.displayName}
-            </option>
-          ))}
-          <option value="none">None</option>
-        </select>
-      </div>
-    );
-  }
+        ))}
+        <option value="none">None</option>
+      </select>
+    </div>
+  );
 }
 
 Selector.propTypes = {
@@ -34,5 +32,3 @@ Selector.propTypes = {
   shelves: PropTypes.array.isRequired,
   onUpdateBook: PropTypes.func.isRequired,
 };
-
-export default Selector;
